+++ webapp/packages/tf/index.ts
@@ -1,3 +1,3 @@
-export * from "./constants";
-export * from "./faceDetectionUtils";
-export * from "./utils";
+export * from "./src/constants";
+export * from "./src/faceDetectionUtils";
+export * from "./src/utils";
