+++ webapp/packages/visualize/src/CanvasElement.tsx
