+++ webapp/src/pages/faceDetection/FaceDetectionPage.tsx
@@ -1,7 +1,7 @@
 import "./faceDetectionPage.scss";
 
-import { runFaceDetecter, STATE, VIDEO_SIZE } from "@js/tf";
-import { CanvasElement, VideoElement } from "@js/visualize";
+import { runFaceDetecter, STATE, VIDEO_SIZE } from "@jj/tf";
+import { CanvasElement, VideoElement } from "@jj/visualize";
 import React, { type FC, useEffect, useState } from "react";
 
 import { Nav } from "../../nav/Nav";
